size_t to int

+++ module3/homework/vectorOfSharedPointers/vectorFunctions.cpp
@@ -6,7 +6,7 @@ std::vector<std::shared_ptr<int>> generate(int count)
 {
     std::vector<std::shared_ptr<int>> vec;
     vec.reserve(count);
-    for (size_t i = 0; i < count; ++i)
+    for (int i = 0; i < count; ++i)
     {
         vec.push_back(std::make_shared<int>(i));
     }
@@ -51,4 +51,3 @@ void sub10(const std::vector<std::shared_ptr<int>>& vec)
         }
     }
 }
-
